+++ src/main/kotlin/org/wycliffeassociates/otter/jvm/app/ui/resources/viewmodel/ResourceListViewModel.kt
@@ -2,7 +2,7 @@ package org.wycliffeassociates.otter.jvm.app.ui.resources.viewmodel
 
 import org.wycliffeassociates.otter.common.data.workbook.*
 import org.wycliffeassociates.otter.common.utils.mapNotNull
-import org.wycliffeassociates.otter.jvm.app.ui.resourcetakes.viewmodel.RecordResourceViewModel
+import org.wycliffeassociates.otter.jvm.app.ui.resourcetakes.viewmodel.ResourceTabPaneViewModel
 import org.wycliffeassociates.otter.jvm.app.ui.workbook.viewmodel.WorkbookViewModel
 import org.wycliffeassociates.otter.jvm.app.widgets.resourcecard.model.ResourceGroupCardItemList
 import org.wycliffeassociates.otter.jvm.app.widgets.resourcecard.model.resourceGroupCardItem
@@ -11,7 +11,7 @@ import org.wycliffeassociates.otter.jvm.utils.onChangeAndDoNow
 import tornadofx.*
 
 class ResourceListViewModel : ViewModel() {
-    internal val recordResourceViewModel: RecordResourceViewModel by inject()
+    internal val resourceTabPaneViewModel: ResourceTabPaneViewModel by inject()
     private val workbookViewModel: WorkbookViewModel by inject()
 
     val resourceGroupCardItemList: ResourceGroupCardItemList = ResourceGroupCardItemList()
@@ -47,7 +47,7 @@ class ResourceListViewModel : ViewModel() {
     internal fun navigateToTakesPage(bookElement: BookElement, resource: Resource) {
         // TODO use navigator to navigate to takes page
         workbookViewModel.activeChunkProperty.set(bookElement as? Chunk)
-        recordResourceViewModel.setRecordableListItems(
+        resourceTabPaneViewModel.setRecordableListItems(
             listOfNotNull(resource.title, resource.body)
         )
     }
