+++ src/main/kotlin/org/wycliffeassociates/otter/jvm/app/ui/resourcetakes/viewmodel/ResourceTabPaneViewModel.kt
@@ -13,7 +13,7 @@ import javafx.collections.ListChangeListener
 import org.wycliffeassociates.otter.jvm.utils.onChangeAndDoNow
 import tornadofx.*
 
-class RecordResourceViewModel : ViewModel() {
+class ResourceTabPaneViewModel : ViewModel() {
     private val workbookViewModel: WorkbookViewModel by inject()
     private val audioPluginViewModel: AudioPluginViewModel by inject()
 
@@ -21,8 +21,8 @@ class RecordResourceViewModel : ViewModel() {
 
     internal val recordableList: ObservableList<Recordable> = FXCollections.observableArrayList()
 
-    class ContentTypeToViewModelMap(map: Map<ContentType, TabRecordableViewModel>):
-        EnumMap<ContentType, TabRecordableViewModel>(map)
+    class ContentTypeToViewModelMap(map: Map<ContentType, RecordableTabViewModel>):
+        EnumMap<ContentType, RecordableTabViewModel>(map)
     val contentTypeToViewModelMap = ContentTypeToViewModelMap(
         hashMapOf(
             ContentType.TITLE to tabRecordableViewModel(),
@@ -31,7 +31,7 @@ class RecordResourceViewModel : ViewModel() {
     )
 
     private fun tabRecordableViewModel() =
-        TabRecordableViewModel(SimpleStringProperty(), audioPluginViewModel)
+        RecordableTabViewModel(SimpleStringProperty(), audioPluginViewModel)
 
     init {
         initTabs()
