+++ src/test/kotlin/org/wycliffeassociates/otter/jvm/app/ui/resources/viewmodel/ResourceListViewModelTest.kt
@@ -8,14 +8,14 @@ import org.junit.Test
 import org.wycliffeassociates.otter.common.data.model.ContentType
 import org.wycliffeassociates.otter.common.data.model.MimeType
 import org.wycliffeassociates.otter.common.data.workbook.*
-import org.wycliffeassociates.otter.jvm.app.ui.resourcetakes.viewmodel.RecordResourceViewModel
+import org.wycliffeassociates.otter.jvm.app.ui.resourcetakes.viewmodel.ResourceTabPaneViewModel
 import org.wycliffeassociates.otter.jvm.app.ui.workbook.viewmodel.WorkbookViewModel
 import tornadofx.*
 
 class ResourceListViewModelTest : ViewModel() {
     private val resourceListViewModel: ResourceListViewModel by inject()
     private val workbookViewModel: WorkbookViewModel by inject()
-    private val recordResourceViewModel: RecordResourceViewModel by inject()
+    private val resourceTabPaneViewModel: ResourceTabPaneViewModel by inject()
 
     init {
         workbookViewModel.activeResourceSlugProperty.set("tn")
@@ -30,9 +30,9 @@ class ResourceListViewModelTest : ViewModel() {
 
     @Test
     fun navigateToTakesPage_callsSetRecordableListItems() {
-        val spiedRecordResourceViewModel = spy(recordResourceViewModel)
+        val spiedRecordResourceViewModel = spy(resourceTabPaneViewModel)
         val spiedResourcesViewModel = spy(resourceListViewModel)
-        whenever(spiedResourcesViewModel.recordResourceViewModel).thenReturn(spiedRecordResourceViewModel)
+        whenever(spiedResourcesViewModel.resourceTabPaneViewModel).thenReturn(spiedRecordResourceViewModel)
 
         // Resource with just a title
         spiedResourcesViewModel.navigateToTakesPage(chunk1, testResourceNoBody)
